Commit 12

+++ src/br/com/pauloAlves_felipeAntonio/projeto_fbd/complemento/TratadorDeMascara.java
@@ -51,7 +51,7 @@ public class TratadorDeMascara {
 		// converte o i-esimo caractere do CPF em um numero:
 		// por exemplo, transforma o caractere '0' no inteiro 0         
 		// (48 eh a posicao de '0' na tabela ASCII)         
-		        num = (int)(CPF.charAt(i) - 48); 
+		        num = CPF.charAt(i) - 48; 
 		        sm = sm + (num * peso);
 		        peso = peso - 1;
 		      }
@@ -65,7 +65,7 @@ public class TratadorDeMascara {
 		      sm = 0;
 		      peso = 11;
 		      for(i=0; i<10; i++) {
-		        num = (int)(CPF.charAt(i) - 48);
+		        num = CPF.charAt(i) - 48;
 		        sm = sm + (num * peso);
 		        peso = peso - 1;
 		      }
